server/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
motor
pymongo
# --- LangChain Core ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

# --- LangChain LLM Providers ---
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
langchain-groq>=0.2.0
groq>=0.9.0
# langchain-upstage>=0.2.0  <-- Removed

# --- Local Embeddings ---
sentence-transformers

# --- LangGraph (Agent Framework) ---
langgraph>=0.2.0

# --- Vector Store ---
langchain-mongodb>=0.1.0

# --- Utils ---
requests
httpx
jinja2
